+++ packages/theme/src/lib/light-theme.ts
@@ -410,6 +410,19 @@ export const lightTheme = {
         maxHeight: '100px',
       },
     },
+    MuiCheckbox: {
+      defaultProps: {
+        color: 'primary',
+      },
+      styleOverrides: {
+        root: {
+          padding: '8px',
+          '&:hover': {
+            backgroundColor: '#00000008',
+          },
+        },
+      },
+    },
     MuiChip: {
       defaultProps: {
         color: 'default',
@@ -500,6 +513,17 @@ export const lightTheme = {
         },
       },
     },
+    MuiFormGroup: {
+      styleOverrides: {
+        root: {
+          '&.MuiFormGroup-row': {
+            '.MuiBox-root:not(:last-of-type)': {
+              marginRight: '16px',
+            },
+          },
+        },
+      },
+    },
     MuiFormHelperText: {
       styleOverrides: {
         root: {
@@ -508,9 +532,6 @@ export const lightTheme = {
       },
     },
     MuiFormLabel: {
-      defaultProps: {
-        shrink: false,
-      },
       styleOverrides: {
         root: {
           // move required asterisk before text
@@ -540,7 +561,7 @@ export const lightTheme = {
       },
       styleOverrides: {
         colorSecondary: {
-          border: `1px solid ${tokens.borderSecondary}`,
+          borderColor: tokens.borderSecondary,
           '&:hover': {
             backgroundColor: tokens.colorActionHover,
           },
@@ -549,7 +570,7 @@ export const lightTheme = {
           },
         },
         colorPrimary: {
-          border: `1px solid ${tokens.borderPrimary}`,
+          borderColor: tokens.borderPrimary,
           '&:hover, &:focus': {
             backgroundColor: tokens.colorActionSelected,
           },