+++ packages/link/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+## [0.2.0](https://github.com/Availity/element/compare/@availity/mui-link@0.1.1...@availity/mui-link@0.2.0) (2023-10-03)
+
+### Features
+
+- **mui-link:** restore system properties on component, remove system properties from docs ([bf864ba](https://github.com/Availity/element/commit/bf864babcb61486da55923af65038aeadce7781e))
+
 ## [0.1.1](https://github.com/Availity/element/compare/@availity/mui-link@0.1.0...@availity/mui-link@0.1.1) (2023-08-23)
 
 ## 0.1.0 (2023-06-05)