+++ packages/typography/CHANGELOG.md
@@ -2,9 +2,10 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
-## 0.1.0 (2023-08-29)
+## [0.1.1](https://github.com/Availity/element/compare/@availity/mui-typography@0.1.0...@availity/mui-typography@0.1.1) (2023-10-03)
 
+## 0.1.0 (2023-08-29)
 
 ### Features
 
-* **mui-typography:** add typography package ([4353d16](https://github.com/Availity/element/commit/4353d162749794063c816afb0bad06a0ae96105f))
+- **mui-typography:** add typography package ([4353d16](https://github.com/Availity/element/commit/4353d162749794063c816afb0bad06a0ae96105f))